driver: workram shared ptr is 32-bit (matches the 32-bit MN10300 bus)
Running the machine reported 'Shared ptr workram found but is width 32, not 8'.
Changed required_shared_ptr<uint8_t> -> <uint32_t>. The kn7000 machine now BOOTS
and RUNS: 'mame kn7000' executes the MN10300 core on the real firmware headless
(~1700% speed, no fatal errors). First time the driver + CPU core actually run.

+++ src/mame/matsushita/kn7000.cpp
@@ -111,7 +111,7 @@ protected:
 private:
 	required_device<mn10300_device> m_maincpu;
 	required_device<screen_device> m_screen;
-	required_shared_ptr<uint8_t> m_workram;
+	required_shared_ptr<uint32_t> m_workram;
 
 	// Control panel button ports and LEDs (CPL = 8 cols, CPC = 5 cols; CPR + the
 	// serial HLE device that reads these / drives the LEDs are still to come).
